+++ src/constants/constants.ts
@@ -1,0 +1,6 @@
+const CONSTANTS = {
+    MS_GRAPH_PATH_USERS:"https://graph.microsoft.com/v1.0/users",
+    MS_GRAPH_PATH_EVENTS: "https://graph.microsoft.com/v1.0/events",
+}
+
+export default CONSTANTS;
