+++ src/example.env
@@ -1,0 +1,1 @@
+AUTH_KEY = "Your auth key from MS Graph Azure";
