+++ src/service/msgraph/participants.service.ts
@@ -1,10 +1,16 @@
-import { Participant } from "../../models/msgraph/participant.entity";
-import { ParticipantRepository } from "../../repository/msgraph/participant.repository";
-
-const participantRepository = new ParticipantRepository();
+import axios from "axios";
+import CONSTANTS from "../../constants/constants";
 
 export class ParticipantService {
-  async getAllParticipants(): Promise<Participant[]> {
-    return participantRepository.getAllParticipants();
+  async getAllParticipants() {
+    try {
+      const participants = await axios.get(
+        CONSTANTS.MS_GRAPH_PATH_USERS,
+        { headers: { Authorization: process.env.AUTH_KEY } }
+      );
+      return participants ? participants.data : null;
+    } catch (err) {
+      console.error(err);
+    }
   }
 }
